Create processing file for serial communication and gamification
This pre file using processing (Java) realises the serial communication from a Brain computer interface called TGAM developed by Neurosky as open-sourced neurofeedback collector.

The project "Metattention" is one of my school work at UCL Computer Science Disability design innovation (collaborated with Global Disability innovation hub). It is originally intended for attention management for people with ADHD, with the extended capability of a larger audience who want to manage their focus. This virtual body-doubling (peer-to-peer competition motivated) system is based on the attentional value gathered with a head-band with TGAM in.

It then uses the attention and meditation value to move two virtual players on the interface visuals. The higher one of the users in this body doubling session, the running speed of the player on visuals (non app) will be higher.

The project remains on the stage of communicating BCI and the Processing visualisation, without further development of front-end development.